refactor to fadecandy_driver / fadecandy_msgs
- Separated msgs to *_msgs package
- Separeted driver to *_driver package
- PEP8 styling refactor
- Added info diagnostics with serial number
