+++ fadecandy_driver/CMakeLists.txt
@@ -1,0 +1,13 @@
+cmake_minimum_required(VERSION 2.8.3)
+project(fadecandy_driver)
+
+find_package(catkin REQUIRED)
+
+catkin_python_setup()
+
+catkin_package()
+
+install(PROGRAMS
+  scripts/fadecandy_node
+  DESTINATION ${CATKIN_PACKAGE_BIN_DESTINATION}
+)
