+++ fadecandy_driver/scripts/fadecandy_node
@@ -1,7 +1,8 @@
 #!/usr/bin/env python
+
 import rospy
 
-from fadecandy_ros.fadecandy_node import FadecandyNode
+from fadecandy_driver.fadecandy_node import FadecandyNode
 
 
 def main():
@@ -10,6 +11,6 @@ def main():
     node = FadecandyNode()
     rospy.spin()
 
+
 if __name__ == '__main__':
     main()
-
