+++ fadecandy_driver/setup.py
@@ -2,7 +2,7 @@ from distutils.core import setup
 from catkin_pkg.python_setup import generate_distutils_setup
 
 d = generate_distutils_setup(
-    packages=['fadecandy_ros'],
+    packages=['fadecandy_driver'],
     package_dir={'': 'src'}
 )
