+++ fadecandy_driver/src/fadecandy_driver/fadecandy_util.py
@@ -1,13 +1,14 @@
+from fadecandy_msgs.msg import LEDStrip
 from std_msgs.msg import ColorRGBA
 
-from fadecandy_ros.msg import LEDStrip
 
 def clamp(v, v_min, v_max):
     return max(v_min, min(v_max, v))
 
+
 def array_to_strip(color_array):
     """
-    Convenience function for converting a 3xN numpy array into an LEDStrip message.
+    Convenience function for converting a 3xN array into an LEDStrip message.
     """
     led_strip = LEDStrip()
     for r, g, b in color_array:
