+++ fadecandy_driver/test/example_client
@@ -1,8 +1,8 @@
 #!/usr/bin/env python
-import rospy
-from std_msgs.msg import ColorRGBA
 
-from fadecandy_ros.msg import LEDArray, LEDStrip
+import rospy
+from fadecandy_msgs.msg import LEDArray, LEDStrip
+from std_msgs.msg import ColorRGBA
 
 
 def main():
@@ -11,10 +11,10 @@ def main():
     led_array_pub = rospy.Publisher('set_leds', LEDArray, queue_size=1)
 
     colors = [
-            ColorRGBA(255, 255, 255, 255),
-            ColorRGBA(255, 0, 0, 255),
-            ColorRGBA(0, 255, 0, 255),
-            ColorRGBA(0, 0, 255, 0)]
+        ColorRGBA(255, 255, 255, 255),
+        ColorRGBA(255, 0, 0, 255),
+        ColorRGBA(0, 255, 0, 255),
+        ColorRGBA(0, 0, 255, 0)]
 
     num_led_strips = 8
     bar_start = 0
@@ -38,6 +38,6 @@ def main():
             color_i = (color_i + 1) % len(colors)
         rospy.sleep(0.1)
 
+
 if __name__ == '__main__':
     main()
-
