+++ fadecandy_driver/udev/10-fadecandy.rules
