+++ fadecandy_msgs/README.md
@@ -1,0 +1,3 @@
+# Fadecandy msgs
+
+ROS msgs for fadecandy LED controllers
