+++ fadecandy_msgs/msg/LEDArray.msg
@@ -1,4 +1,4 @@
 # The fadecandy board supports up to 8 LED strips. Each strip can have up to
 # 64 LEDs.
-fadecandy_ros/LEDStrip[] strips
+LEDStrip[] strips
