+++ fadecandy_msgs/msg/LEDStrip.msg
